Add security token for deletion pin

+++ src/Controller/PinsController.php
@@ -72,10 +72,12 @@ class PinsController extends AbstractController
     }
 
     #[Route('/pins/{id}/delete', name: 'app_pins_delete', methods: ['DELETE'])]
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pin_delete_' . $pin->getId(), $request->request->get('_csrf_token'))) {
+            $em->remove($pin);
+            $em->flush();
+        }
 
         return $this->redirectToRoute('app_home');
     }
